externalize storable decoding
