+++ cbor_bench_test.go
@@ -59,14 +59,14 @@ func benchmarkEncodeCBORArray(b *testing.B, values []Storable) {
 		var buf bytes.Buffer
 		enc := newEncoder(&buf)
 
-		enc.scratch[0] = 0x80 | 27
-		binary.BigEndian.PutUint64(enc.scratch[1:], uint64(len(values)))
-		enc.Write(enc.scratch[:9])
+		enc.Scratch[0] = 0x80 | 27
+		binary.BigEndian.PutUint64(enc.Scratch[1:], uint64(len(values)))
+		enc.Write(enc.Scratch[:9])
 
 		for _, v := range values {
 			v.Encode(enc)
 		}
-		enc.cbor.Flush()
+		enc.CBOR.Flush()
 	}
 }
 
@@ -145,13 +145,13 @@ func encode(values []Storable) []byte {
 	var buf bytes.Buffer
 	enc := newEncoder(&buf)
 
-	enc.scratch[0] = 0x80 | 27
-	binary.BigEndian.PutUint64(enc.scratch[1:], uint64(len(values)))
-	enc.Write(enc.scratch[:9])
+	enc.Scratch[0] = 0x80 | 27
+	binary.BigEndian.PutUint64(enc.Scratch[1:], uint64(len(values)))
+	enc.Write(enc.Scratch[:9])
 
 	for _, v := range values {
 		v.Encode(enc)
 	}
-	enc.cbor.Flush()
+	enc.CBOR.Flush()
 	return buf.Bytes()
 }
